+++ Model/Payment/Paybuttons.php
@@ -27,7 +27,6 @@ class Paybuttons extends AbstractPayment
     const XML_PATH = 'payment/pbxep_paybuttons/cctypes';
 
     protected $_code = self::CODE;
-    protected $_3dsAllowed = true;
     protected $_hasCctypes = true;
     protected $_allowManualDebit = true;
     protected $_allowDeferredDebit = true;