+++ Model/Payment/PrivateCard.php
@@ -27,7 +27,6 @@ class PrivateCard extends AbstractPayment
     const XML_PATH = 'payment/pbxep_private/cctypes';
 
     protected $_code = self::CODE;
-    protected $_3dsAllowed = true;
     protected $_hasCctypes = true;
     protected $_allowManualDebit = true;
     protected $_allowDeferredDebit = true;