+++ Model/Ui/PbxepcbConfigProvider.php
@@ -63,7 +63,7 @@ final class PbxepcbConfigProvider implements ConfigProviderInterface
             $types = explode(',', $types);
         }
         foreach ($types as $code) {
-            $asset = $assetRepository->createAsset('Paybox_Epayment::images/' . strtolower($code) . '.45.png');
+            $asset = $assetRepository->createAsset('Paybox_Epayment::images/' . strtoupper($code) . '.svg');
             $placeholder = $assetSource->findRelativeSourceFilePath($asset);
             if ($placeholder) {
                 list($width, $height) = getimagesize($asset->getSourceFile());