+++ Model/Ui/PbxepprepaidConfig.php
@@ -46,7 +46,6 @@ class PbxepprepaidConfig
      */
     protected $_isOffline = false;
     protected $scopeConfig;
-    protected $_3dsAllowed = true;
     protected $_hasCctypes = true;
     protected $_allowManualDebit = true;
     protected $_allowDeferredDebit = true;