+++ Model/Ui/PbxepprivateConfig.php
@@ -56,7 +56,6 @@ class PbxepprivateConfig
      */
     protected $_isOffline = false;
     protected $scopeConfig;
-    protected $_3dsAllowed = true;
     protected $_hasCctypes = true;
     protected $_allowManualDebit = true;
     protected $_allowDeferredDebit = true;